# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface FoodItem {
  id: string;
  name: string;
  description: string;
  price: number;
  category: string;
  image: string;
  isVeg: boolean;
  rating: number;
  ratingCount: number;
  isAvailable: boolean;
  isPopular: boolean;
  preparationTime: number;
}

export interface PickupSlot {
  id: string;
  startTime: string;
  endTime: string;
  capacity: number;
  booked: number;
  isAvailable: boolean;
}

export interface OrderItem {
  id: string;
  name: string;
  price: number;
  quantity: number;
  image: string;
}

export type OrderStatus = typeof OrderStatus[keyof typeof OrderStatus];


export const OrderStatus = {
  placed: 'placed',
  accepted: 'accepted',
  preparing: 'preparing',
  ready: 'ready',
  completed: 'completed',
} as const;

export interface Order {
  id: string;
  items: OrderItem[];
  total: number;
  status: OrderStatus;
  pickupTime: string;
  pickupToken: string;
  placedAt: string;
}

export type OrderInputPaymentMethod = typeof OrderInputPaymentMethod[keyof typeof OrderInputPaymentMethod];


export const OrderInputPaymentMethod = {
  upi: 'upi',
  cash: 'cash',
} as const;

export interface OrderInput {
  items: OrderItem[];
  pickupSlotId: string;
  paymentMethod: OrderInputPaymentMethod;
}

export interface Notification {
  id: string;
  type: string;
  title: string;
  message: string;
  time: string;
  read: boolean;
}