# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /orders:
    get:
      operationId: getOrders
      tags: [orders]
      summary: Get current student's orders
      responses:
        "200":
          description: Student orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Place a canteen order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Created order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{orderId}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get one order
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Order details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Order not found
...
components:
  schemas:
...
    OrderItem:
      type: object
      required: [id, name, price, quantity, image]
      properties:
        id: { type: string }
        name: { type: string }
        price: { type: number }
        quantity: { type: number }
        image: { type: string }
    Order:
      type: object
      required: [id, items, total, status, pickupTime, pickupToken, placedAt]
      properties:
        id: { type: string }
        items:
          type: array
          items: { $ref: "#/components/schemas/OrderItem" }
        total: { type: number }
        status:
          type: string
          enum: [placed, accepted, preparing, ready, completed]
        pickupTime: { type: string }
        pickupToken: { type: string }
        placedAt: { type: string }
    OrderInput:
      type: object
      required: [items, pickupSlotId, paymentMethod]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/OrderItem" }
        pickupSlotId: { type: string }
        paymentMethod: { type: string, enum: [upi, cash] }
...