# lib/api-zod/src/generated/types/orderInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { OrderInputPaymentMethod } from './orderInputPaymentMethod';
import type { OrderItem } from './orderItem';

export interface OrderInput {
  items: OrderItem[];
  pickupSlotId: string;
  paymentMethod: OrderInputPaymentMethod;
}